SecureDataFetchMPC/requirements.txt
requests==2.31.0
cryptography==41.0.5
openpyxl==3.1.2
pandas==2.0.3
tqdm==4.66.1
python-dotenv==1.0.0